requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.2
azure-cosmos==4.7.0
azure-storage-blob==12.23.1
azure-identity==1.18.0
azure-servicebus==7.12.3
bandit==1.7.9
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
coverage==7.5.4
dill==0.3.8
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
orjson==3.10.5
packaging==24.1
pbr==6.0.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-pool==3.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pylint==3.2.3
pytest==8.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
requests==2.32.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
stevedore==5.2.0
tomlkit==0.12.5
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
